Add inherent playlist_id and playlist_item_id as properties for the Content  object

+++ blynq/contentManagement/models.py
@@ -282,13 +282,41 @@ class Content(models.Model):
             debugFileLog.exception("Exception while saving the playlist to update relevant schedules")
             mail_exception(exception=e)
 
-    def delete_user_invisible_playlist(self):
+    def get_user_invisible_playlist_item(self):
         try:
+            self.playlist_item
+        except AttributeError:
+            self.playlist_item = None
+        if self.playlist_item:
+            return self.playlist_item
+        else:
             from playlistManagement.models import PlaylistItems
             playlist_items = PlaylistItems.objects.select_related('playlist').filter(content_id=self.content_id,
-                                                                                     playlist__user_visible=False)
-            for item in playlist_items:
-                playlist = item.playlist
+                                                                                    playlist__user_visible=False)
+            if playlist_items.exists():
+                return playlist_items[0]
+            else:
+                return None
+
+    @property
+    def playlist_id(self):
+        # This indicates the playlist_id of user_invisible playlist which is created for every content except folder
+        playlist_item = self.get_user_invisible_playlist_item()
+        playlist_id = playlist_item.playlist_id if playlist_item else None
+        return playlist_id
+
+    @property
+    def playlist_item_id(self):
+        # This indicates playlist_item_id of user_invisible playlist which is created for every content except folder
+        playlist_item = self.get_user_invisible_playlist_item()
+        playlist_item_id = playlist_item.playlist_item_id if playlist_item else None
+        return playlist_item_id
+
+    def delete_user_invisible_playlist(self):
+        try:
+            playlist_item = self.get_user_invisible_playlist_item()
+            if playlist_item:
+                playlist = playlist_item.playlist
                 playlist.delete()
         except Exception as e:
             mail_exception(exception=e)
@@ -321,12 +349,10 @@ class Content(models.Model):
         try:
             if self.is_folder:
                 return
-            from playlistManagement.models import PlaylistItems
-            playlist_item = PlaylistItems.objects.select_related('playlist').filter(content_id=self.content_id,
-                                                                                    playlist__user_visible=False)
-            if playlist_item.exists():
+            playlist_item = self.get_user_invisible_playlist_item()
+            if playlist_item:
                 # Update user invisible playlist title
-                playlist = playlist_item[0].playlist
+                playlist = playlist_item.playlist
                 playlist.playlist_title = self.title
                 # Below line not required after initial merge
                 playlist.playlist_type = self.get_playlist_type()
